Added an APIRandomUser to load the raw data loaded from the API service response. Inversion of control for the api worker dependency to be configured at time of interactor creation.
Created the client builder to connect to a service. Created a api call builder to make requests to the RandomUsersApi. Implemented the RandomUserAPIWorker that downloads the random users from the api and feeds the interactor with the results.

Connected the RandomUserListInteractor to use the api worker in order to download real data.

Solved bug to the filter, to be capitalization adnostic.